==> analysis/1.get-data-raw.ipynb <==
# Import libraries

import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
from IPython.core.debugger import set_trace
import json
from matplotlib import colors
from ast import literal_eval

# Load data

# psynet 11
fname = "sing-scales-v1"
fname = "sing-scales-v2"

PATH = f"data/{fname}/"
nodes = pd.read_csv(PATH + "CustomNode.csv", low_memory=False)
# networks = pd.read_csv(PATH + "CustomNetwork.csv", low_memory=False) # old psynet
trials = pd.read_csv(PATH + "CustomTrial.csv", low_memory=False)

print(nodes)

# filter nodes
node_data = nodes
node_data = node_data[node_data["failed"] == False]
node_data = node_data[["id", "network_id", "degree", "trial_maker_id", "definition", "seed"]]
node_data = node_data.sort_values(["network_id", "degree"])
node_data["definition"].iloc[1]

def find_end_tt(d, end_tt_list=None):
    if end_tt_list is None:
        end_tt_list = []
    if isinstance(d, dict):
        for key, value in d.items():
            if key == "start_tt":
                end_tt_list.append(value)
            elif isinstance(value, (dict, list)):
                find_end_tt(value, end_tt_list)
    elif isinstance(d, list):
        for item in d:
            find_end_tt(item, end_tt_list)
    return end_tt_list


node_data['start_tt'] = node_data['definition'].apply(lambda x: find_end_tt(json.loads(x)))

print(node_data['start_tt'])


# Preprocess data

# filter nodes
node_data = nodes
node_data = node_data[node_data["failed"] == False]
node_data = node_data[["id", "network_id", "degree", "trial_maker_id", "definition", "seed"]]
node_data = node_data.sort_values(["network_id", "degree"])

# process node definition
mel_id = []
# register = []
trial_type = []
target_pitches = []
# reference_pitch = []
target_intervals = []
target_note_durations = []
# target_ISIs = []

for i in range(len(node_data)):
    print()
    # definition = node_data["definition"].to_numpy()[i]
    # definition = literal_eval(definition)
    # definition = [d["content"] for d in definition if d["is_center"]][0]

    definition_str = node_data["definition"].iloc[i]  # Access the string directly
    print(definition_str)
    definition = json.loads(definition_str)  # Parse JSON string into dictionary
    # definition = [d["content"] for d in definition if d["is_center"]][0]
    
    # mel_id.append(definition["mel_id"])
    # register.append(definition["register"])
    trial_type.append(definition["trial_type"])
    target_pitches.append(definition["target_pitches"])
    # reference_pitch.append(definition["reference_pitch"])
    target_intervals.append(definition["target_intervals"])
    target_note_durations.append(definition["target_note_durations"])
    # target_ISIs.append(definition["target_ISIs"])

# node_data["mel_id"] = mel_id
# node_data["register"] = register
node_data["trial_type"] = trial_type
node_data["target_pitches"] = target_pitches
# node_data["reference_pitch"] = reference_pitch
node_data["target_intervals"] = target_intervals
node_data["target_note_durations"] = target_note_durations
# node_data["target_ISIs"] = target_ISIs

node_data

# Export node data to csv

node_data.to_csv(f"{fname}_node_data.csv")

# Trial level data

# filter trials
trial_data = trials
trial_data = trial_data[trial_data["failed"] == False]

degrees = []

for i in range(len(trial_data)):
    current_node_id = trial_data.iloc[i]["origin_id"]
    current_nodes = node_data[node_data["id"] == current_node_id]
    current_degree = current_nodes["degree"].to_numpy()[0]
    # answer = d['listen_and_rate_trial_seed']['valence']
    # current_vertex_id = current_nodes["vertex_id"].to_numpy()[0]
    # current_dependent_vertex_ids = current_nodes["dependent_vertex_ids"].to_numpy()[0]
    degrees.append(current_degree)
    # vertex_ids.append(current_vertex_id)
    # dependent_vertex_ids.append(current_dependent_vertex_ids)

trial_data["degree"] = degrees

trial_data = trial_data[["id", "network_id", "degree", "trial_maker_id", "origin_id", "participant_id", "analysis"]]
trial_data = trial_data.sort_values(["network_id", "degree"])

trial_data

# Export trial data to csv

trial_data.to_csv(f"{fname}_trial_data.csv")

# Experiment Cost and Demographics

participants = pd.read_csv(PATH + "participant.csv", low_memory=False)

valid_participants = participants[participants["complete"] == True]
valid_participants = valid_participants[valid_participants["failed"] == False]
experiment_summary = {
    "N_participants": valid_participants.shape[0],
    "cost": participants["base_pay"].sum() + participants["bonus"].sum()
}

experiment_summary
